+++ src/Game/Segments/Overlay_22.cs
@@ -974,8 +974,8 @@ namespace OpenCiv1
 			// Instruction address 0x0000:0x0b79, size: 5
 			this.oParent.Segment_1d12.F0_1d12_0045(this.oCPU.ReadInt16(this.oCPU.SS.Word, (ushort)(this.oCPU.BP.Word - 0x6)), -1);
 
-			this.oCPU.AX.Word = this.oParent.Var_70e2;
-			this.oCPU.AX.Word = this.oCPU.SUBWord(this.oCPU.AX.Word, this.oParent.Var_70e4);
+			this.oCPU.AX.Word = (ushort)((short)this.oParent.Var_70e2);
+			this.oCPU.AX.Word = this.oCPU.SUBWord(this.oCPU.AX.Word, (ushort)((short)this.oParent.Var_70e4));
 			this.oCPU.CX.Low = 0x5;
 			this.oCPU.AX.Word = this.oCPU.SHLWord(this.oCPU.AX.Word, this.oCPU.CX.Low);
 			this.oCPU.CWD(this.oCPU.AX, this.oCPU.DX);
